renamed spatial_proteomics to spatialproteomics
