+++ spatialproteomics/constants.py
