+++ tests/pp/test_grow_cells.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from spatial_proteomics.constants import Dims, Layers
+from spatialproteomics.constants import Dims, Layers
 
 
 def test_grow_cells(dataset):
