+++ tests/pp/test_handle_disconnected_cells.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from spatial_proteomics.pp.utils import handle_disconnected_cells
+from spatialproteomics.pp.utils import handle_disconnected_cells
 
 
 def test_handle_disconnected_cells():
