+++ tests/pp/test_slicing.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from spatial_proteomics.constants import Dims, Features, Layers
+from spatialproteomics.constants import Dims, Features, Layers
 
 
 def test_image_slicing_two_coordinates(dataset):
